requirements.txt
streamlit==1.24.0
tensorflow==2.13.0
opencv-python-headless==4.8.0.76
numpy==1.24.3
pillow==9.5.0
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2